Removed /team, combined with /about and changed placeholder images
